Add favorites model and relations to users and movies
A favorites model, migration file, controller, and views have been added to the application. Relationships between the tables were defined in models as one-to-many: a user can have many favorites, as can a movie. The user and movie show views were updated to reflect these changes: they can now display a user's favorite movies, and the fans of a movie. The routes were also updated to include the new favorites resources.