Implement slug usage for User and Genre models
Introduced slug concept to both User and Genre models that changes the way data is accessed from the id to the more readable slug. This involved creating database migrations to add slug fields, adjusting the controllers' methods for these models, and adding before_save callbacks to format and set slug based on existing data. The commit aims to improve URL readability and maintainability.