Replace 'sprockets-rails' gem with 'dartsass-sprockets' and update dependencies
Switched the gem 'sprockets-rails' with 'dartsass-sprockets' in the Gemfile. Also, the gems for rails, actioncable, actionmailbox, actionmailer, and several other dependencies are updated to the latest versions in Gemfile.lock. These updates are necessary for ensuring that the application benefits from the latest features, security updates, and bug fixes introduced in these new versions.