Remove 'flipper' feature flagging system
This commit completely removes the 'flipper' feature flagging system from the project. All dependencies related to 'flipper' in Gemfile and Gemfile.lock, and all related code in views, routes and schema have been removed. The decision to remove this was based on the lack of use of feature flags in the current project.