Refactor AWS S3 storage settings in Rails app
This commit rearranges the AWS S3 storage configuration in the storage.yml file of the Rails application. It also uses Rails' built-in functionality to set the access_key_id and secret_access_key from credentials. The bucket name for storing files has been set dynamically based on the environment name.